Refactor queries.js
