+++ server/pool.js
@@ -1,0 +1,13 @@
+const Pool = require('pg').Pool
+const dotenv = require(('dotenv'))
+dotenv.config()
+
+const pool = new Pool({
+    user: process.env.PGUSER,
+    host: process.env.PGHOST,
+    database: process.env.PGDATABASE,
+    password: process.env.PGPASSWORD,
+    port: process.env.PGPORT,
+})
+
+module.exports = pool
